requirements.txt
Flask==3.0.2
MarkupSafe==2.1.5
pymongo==4.6.2
flask-cors==3.0.10
python-decouple==3.8
reportlab==4.1.0
